+++ dags/annuaire_entreprise_checks.py
@@ -3,13 +3,13 @@ from datetime import datetime
 
 import pandas as pd
 import utils.base_utils as base_utils
-import utils.dag_eo_utils as dag_eo_utils
 import utils.mapping_utils as mapping_utils
 import utils.siret_control_utils as siret_control_utils
 from airflow import DAG
 from airflow.models.param import Param
 from airflow.operators.python import PythonOperator
 from airflow.providers.postgres.hooks.postgres import PostgresHook
+from tasks.airflow_logic.write_data_task import write_data_task
 
 pd.set_option("display.max_columns", None)
 
@@ -318,11 +318,6 @@ get_location_task = PythonOperator(
     dag=dag,
 )
 
-write_data_task = PythonOperator(
-    task_id="db_data_write",
-    python_callable=dag_eo_utils.write_to_dagruns,
-    dag=dag,
-)
 
 serialize_to_json_task = PythonOperator(
     task_id="db_data_prepare",
@@ -343,5 +338,5 @@ combine_candidates = PythonOperator(
     >> combine_candidates
     >> get_location_task
     >> serialize_to_json_task
-    >> write_data_task
+    >> write_data_task(dag)
 )
